+++ packages/components/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.6.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.5.4...@jobber/components@6.6.0) (2024-12-04)
+
+
+### Features
+
+* **components:** Add controlled option to Tabs ([#2204](https://github.com/GetJobber/atlantis/issues/2204)) ([2c7559c](https://github.com/GetJobber/atlantis/commit/2c7559c98eb947b163e1fc80d329052caa59c34e))
+
+
+
+
+
+## [6.5.4](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.5.3...@jobber/components@6.5.4) (2024-12-04)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [6.5.3](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.5.2...@jobber/components@6.5.3) (2024-11-29)